Use /var/data for ZipDownloadJob tmp dir when available
Render allows worker machines to have additional disks mounted to
them. We saw an issue in production where an artist attempted to
upload a large file (around 1GB) and the ZipDownloadJob failed due to
a lack of disk space.

This commit means that the ZipDownloadJob checks for the existance of
the `/var/data` path and then uses that as the root for the tmpdir if
it exists.

To use this in production we need to add a disk to the sidekiq worker
machine and mount it at this path.