Use /var/data for TranscodeJob tmp dir when available
Render allows worker machines to have additional disks mounted to
them. We saw an issue in production where an artist attempted to
upload a large file (around 1GB) and the ZipDownloadJob failed due to
a lack of disk space. We fixed that in the previous commit. A similar
thing could potentially happen with this job too, so it makes sense to
also use any additional disk space we have.