Record amount of tax added to a purchase when checkout successful
When running the payouts to artists at the end of the month we need[1] to
know the amount of tax added to each purchase by stripe. We receive
that in the payload of the `checkout.session.completed` event

[1] We're assuming for now, while waiting on advice from our
accountant, that we won't pass any sales tax that's added to the sale
to the artist.