Retry TranscodeJob on DeserializationError
In 8c5de8293 we decided to discard jobs if this exception occurs
because we thought they were only occuring when a track had been
deleted. However we've noticed today that some transcodes were not
being generated at all even for tracks that hadn't been deleted. It
looks like there might be some racy behaviour somewhere that means a
track has not been saved when this job is enqueued. At some point we
might need to track that down, but to improve things for now I'm using
`retry_on` to retry the jobs 5 times (the default) waiting for 10
seconds between each retry. I'll see if that improves things while
hopefully reducing the number of Rollbar notices we get for this
exception.