Only enable track player if file is available
We've seen a couple of exceptions[1] on this view where the `file` for
`track.preview` is `nil`. I think this could happen if the transcode
has not completed, but I haven't looked into it in any detail. I've
added an additional guard around the relevant code in the track
list (and also in a similar location for the audio controls).

[1] https://app.rollbar.com/a/gofreerange/fix/item/jam/12/occurrence/385397023393