Add a payment details form to collect user's name
To make a payment to an artist using Wise we need to know the "full
name of the account holder". This form allows us to collect that via a
link from the settings page.

The call to `resolve` in `config/routes.rb` is required to generate
the corrrect routes in `form_with` helpers given `payout_detail` is a
singular resource[1]

[1] https://guides.rubyonrails.org/form_helpers.html#relying-on-record-identification