Ignore all exceptions handled by Rails ActionDispatch::ShowExceptions
I can't see why we'd want to report any of the exceptions which
ActionDispatch::ShowExceptions rescues and converts into suitable HTTP
response codes using ActionDispatch::ExceptionWrapper [1]. This approach
is probably somewhat draconian and we might miss some important
exceptions. However, I don't think this is a big risk and the benefit is
to improve the signal-to-noise ratio in our exception monitoring.

This commit adds a before_process block to the Rollbar configuration so
that we only report exceptions which will not be handled by the Rails
middleware.

c.f. https://github.com/freerange/site/commit/1bf008051eff4cda8c0029cc855279868610725e

[1]: https://github.com/rails/rails/blob/v7.1.1/actionpack/lib/action_dispatch/middleware/exception_wrapper.rb#L11-L25