Validate cover file is an image
We've seen an exception [1] where someone uploaded a PDF as the album
cover. This adds validation for the album cover so that it must be a
jpeg or a png. This covers all the existing albums, but we might want to
also allow other image types, e.g. gif, bmp, svg, x-png?

I've also added the `accepts` option to the `file_field` for an album
cover which should make it harder (impossible?) to select a file that is
not a jpeg or a png.

I've also added a warning like we have for the tracks files to explain
that the cover must be a jpeg or a png. Apologies in advance if my
tailwind isn't up to scratch for this - I've just copied the markup for
the tracks warning.

I also note that we're not currently requiring a cover image for an
album. I don't know whether this is intentional or an oversight, but
I've left this aspect of the validation unchanged for the moment.

[1]: https://app.rollbar.com/a/gofreerange/fix/item/jam/15