Make Sidekiq web UI available to admin users
This means that when an admin user is signed in, they can view the
Sidekiq dashboard at `/sidekiq`. I think this could be quite useful for
diagnosing problem (and even doing things like killing jobs) while we're
in this early development phase of the project.

The authentication/authorization documentation in this Sidekiq
documentation [1] seems to use some outdated Rails mechanism. So I've
used an adapted version of the code in this article [2] instead.
However, since I'm not very familiar with the authentication mechanism
we're using in this app, I'm not sure what I've done is the simplest
solution, although it does work!

[1]: https://github.com/sidekiq/sidekiq/wiki/Monitoring#authentication
[2]: https://mauromorales.com/2022/08/02/rails-routing-advanced-constraints-for-user-authentication-without-devise/