Allow user to select a payout country/region
Last month we ran the payouts to artists using wise. Wise supports a
number of different countries/regions[1], so I've turned that list
into a helper for the user to select.

I've excluded for now the countries that you can send USD/GBP/EUR to
but are outside of the countries that usually use those
currencies. These countries incur additional charges and we'd probably
need to have a minimum payout threshold before it would make sense to
do that.

[1] https://wise.com/help/articles/2571942/what-countriesregions-can-i-send-to