Display filename of existing cover on album form
c.f. what we're doing for the track original files in the same form.

We might want to try to extract the duplication, but I'm a bit unsure
how to do that with the current `TailwindFormBuilder` implementation. We
could probably do it in a helper method, but that doesn't feel quite
right either.

We might also want to consider doing this for the profile picture file
field on the artist form, but I'm trying to keep this PR focussed on the
album form.