Add track number to ID3 metadata
I've reused the formatted track position that was already being used for
the filename prefix in `ZipDownloadJob#track_filename`, because I
thought it was worth being consistent.

We're using ID3v2.3 and it seems as if the `TRCK` tag is the one we
want. See the earlier commit [1] where we added the other metadata for
more details. As in that commmit, `Track#number` will never be blank, so
there's no need to add it to the metadata hash individually.

I've added a check for a track's position changing to the `after_commit`
callback which trgiggers re-transcoding.

We'll need to re-transcode all tracks in order to apply this to existing
tracks, but I suggest we tackle that as a separate manual task where
we'll need to run `Artist.all.each(&:transcode_albums)` in a Rails
console.

[1]: https://github.com/freerange/music-coop/commit/1851e27b783f24a3ad459d442737e5d34ca807e4