Display filename of existing artist profile picture
Unlike in this commit [1], I've used the built-in
`ActionView::Helpers::FormBuilder#label` method rather than the
customized `TailwindFormBuilder#labels` method that I used for the file
fields in the album form. That's because at the moment there aren't any
validation rules on `Artist#profile_picture`, so we don't (yet) need to
be able to display validation errors withinin the label.

c.f. this commit [1].

See #60.

[1]: https://github.com/freerange/music-coop/commit/3aaec284d7d07355620dd646529ba94bcf2287b8