Reuse alert_box in application layout
When I added the alert markup to the application layout in this earlier
commit [1], I hadn't noticed the `app/views/shared/_alert_box.html.erb`
shared partial. This was being used in pages containing forms but, for
the scenario in [1], we need it to exist in *all* pages. So I've
replaced the custom markup in the application layout with the shared
partial and remove the use of the shared partial from the 5 form pages.

This means a very slight change of position of the markup within
`app/views/sessions/new.html.erb` - previously the alert was inside the
form element - however, I don't this this makes a material difference to
where the alert appears on the page.

Note that the previous implementation meant that two different versions
of the alert were being displayed on some pages. This is now fixed.

[1]: https://github.com/freerange/music-coop/commit/c6ebe0ae11c2eab262b62a6f83d2689a0625c8c7