Reuse notice_box in application layout
Previously the notice message was only being displayed on some pages.
This moves it into the application layout so that it can appear on any
page which makes it more consistent with the way the alert message now
works.

Note that this means a slight change in position in
`app/views/users/show.html.erb` where previously the notice message
appeared *below* the page heading. It's not obvious to me that the
latter was intentional and it seems inconsistent, so I'm going to
change it and hope for the best!