Use after_commit vs after_save to queue TranscodeJob
A while ago we were seeing `ActiveJob::DeserializationError` exceptions
like this one [1]. We tried to workaround it in this commit [2] and then
more successfully in this other commit [3], but we didn't tackle the
root cause.

I suspect the root cause was that we were using an `after_save` callback
on `Track` and other models to enqueue `TranscodeJob`s rather than using
an `after_commit` callback as explained in this blog post [4].

I've also removed the `retry_on` workaround from `TranscodeJob` in the
hope that we no longer need it.

Closes #115.

[1]: https://app.rollbar.com/a/gofreerange/fix/item/jam/11
[2]: https://github.com/freerange/music-coop/commit/8c5de8293cea5837076ec74ebe2aa1a682392d3c
[3]: https://github.com/freerange/music-coop/commit/79f278bb9474ae57e4bece3e3679723f28369af9
[4]: https://www.rubyinrails.com/2018/12/03/rails-usage-of-after-commit-vs-after-save/