Enable client-side exception monitoring with rollbar
We've had a report (#120) of errors uploading tracks which look like
they're being thrown from the client (likely from the direct upload
javascript). To understand a bit more about what's going on this
commit enables client side exception monitoring.

I added the post_client_item access token from our rollbar
dashboard[1] to the development and production credentials file.

It's quite likely that these exceptions will be a bit noisy, but we
can try to ignore ones we can't do anything about, or revisit having
this enabled once we get to the bottom of what's happening with the
track upload issue.

[1] https://app.rollbar.com/a/gofreerange/p/jam/settings/access_tokens