Replace encrypted credentials with environment variables
We eventually want to open up this repository. Despite the encryption
I'm a little nervous including production credentials in a public
repo. Also, if a non-GFR developer wants to run this application
they'll likely want to be able to use their own aws, stripe, postmark
and rollbar services, so having the shared credentials for production
and development makes less sense.

This commit instead uses environment variables (as defined in
`.yml` files under `config`) to store the existing tokens for third party
services. The `.env.example` file shows what values these can have in
development.

I'll remove the existing `config/credentials` in a subsequent commit.

Before this commit can be deployed, we'll need to set these 8
variables in production.

Note, the existing `config/credentials/*.yml.enc` files also define a
value for `secret_key_base`. Rails already[1] uses the value stored in
the `SECRET_KEY_BASE` env var in preference to this, so we just have
to remember to also set that in production. In development the value
of `tmp/local_secret.txt` is used anyway[1] so it doesn't need to be
set in `.env`.

I've also had to set these variables (by copying the `.env.example` to
`.env.test` in the CI environment and ensure it's loaded by the linter
task as well).

I also stubbed out the value of `Rails.configuration.postmark` in a
test as when POSTMARK_WEBHOOKS_TOKEN is set to `nil` the tests
fail (`NoMethodError: undefined method `bytesize' for nil:NilClass` in
EmailSubscriptionChangesController#authenticate)

[1] https://api.rubyonrails.org/classes/Rails/Application.html#method-i-secret_key_base