Only allow track re-ordering for unpublished albums
At the moment the zip files containing all the transcoded tracks for an
album are created in `Album#publish`. However, if we re-order the tracks
in an album, then the metadata for the tracks will change and we'll need
to re-transcode all the tracks and re-create the zip files.

The re-transcoding is already handled by the `after_commit` callback on
`Track`. However, there isn't currently a mechanism to re-create the zip
files. We could introduce such a mechanism using callbacks, but for now
I think it's simplest to only allow track re-ordering for unpublished
albums. That way the zip files will be re-created at the point the album
is published. If a user needs to re-order the tracks for a published
album, they can unpublish it, re-order the tracks, and then re-publish
it.

I wasn't 100% sure that using `TrackPolicy` to enforce this constraint
was appropriate. However, there is some precedence for checking the
publication status of an album in `AlbumPolicy#show?`, so I think it's
probably OK.

Note that we might want to consider preventing users from changing other
data relating to published albums or their tracks in order to ensure the
zip files are always up-to-date.