Remove config/credentials
We're now using environment variables to set access tokens and other
configuration for stripe, rollbar, postmark and aws. I've set the
appropriate variables in Render using the ones stored in
config/credentials/production.yml.enc. For development the app should
boot if `.env.example` is copied to `.env` and we can fetch the latest
values of the variables from those services if we need to use them in
development.