Add PurchaseExporter
We need to run a round of payouts again, I'm hoping this
quick-and-dirty exporter will make the process a bit easier by giving
me the data in a form I can query/group in a spreadsheet.

We need to know the Wise payout details, the price of each sale, and any
tax that was charged to calculate the payout for each artist. I've
added the album title and artist name to the sheet to make it easier
to cross-check with the site itself, or email the artist if we have
any queries.