Remove old credentials for the AWS IAM user
I've updated the relevant env vars in Render to use the new credentials,
so these old credentials are redundant.

In hindsight, I probably should've used the access key serial number
(i.e. 1) in the `CfnAccessKey` construct name rather than the word
"New", because now there's only one access key I think the naming is a
bit confusing. However, I can't easily change it, because CloudFormation
uses the name to identify the underlying resource and thus changing it
would lead to the current access key being removed and a new one being
created. So I'm going to leave it for now.

This is the last step in addressing this request [1] in order that we
remove all live credentials from the git repo.

[1]: https://github.com/freerange/music-coop/issues/107#issuecomment-1878842968