Don't set Purchase#user on purchase complete if already set
Someone contacted me to say they'd made a purchase and it wasn't
appearing in their collection. I think what might have happened is
that they were logged in when they made the purchase and then used a
different email at Stripe checkout and because a user with that email
couldn't be found, Purchase#user was set to nil.

This should fix that.