Add Album#first_published_on
The plan is to use this as a kind of default for `Album#released_at`.
See #65 for some discussion about this.

It's far from perfect, but I've attempted to backfill values for all the
existing published albums by taking the latest of `Album#created_at` and
the latest of `Track#created_at` for all the album's tracks

Initially I was going to use `Album#updated_at` & `Track#updated_at`,
but it looks as if some of our bulk changes (maybe the recent
re-transcoding?) have updated these values such that many of them have
very similar values. Using `Album#created_at` & `Track#created_at` seems
to give more sensible values.