Add AlbumPolicy::Scope & use in ArtistsController#show
I think this is more idiomatic use of Pundit and it simplifies
`ArtistsController#show` nicely which will make adding an Atom feed for
that action easier.

I suspect there's a nicer way to implement the 2nd expression in the
`if` statement in `AlbumPolicy::Scope#resolve`, but the one I've gone
with seems to result in a sensible number of queries.