Add Artist#first_listed_on
Since an artist is considered to be listed when at least one of their
albums is published, I think it makes sense to say
`Artist#first_listed_on` should be the oldest of
`Album#first_published_on` on any of its albums. And that should be
correct even if an album has subsequently been unpublished.

I'm planning to use `Artist#first_listed_on` for the entry.published
property on an Atom feed for `ArtistsController#index`, i.e. for each
artist.