Only capitalize first word in error message on field
Previously we were seeing error messages like this in form field labels:

    File Cannot Be Missing
    Must Be Less Than Or Equal To 2024 01 09

I think the new version is more idiomatic and likely to cope better when
there are multiple errors on a single field.

The new version generates the following error messages for the same
fields:

    Cover file cannot be missing
    Released on must be less than or equal to 2024-01-09

While this duplicates the label name (i.e. "Cover" & "Released on"), I
think it reads better and it matches the error message in the summary at
the top of the page.