Feature purchased albums on home page
This is a bit of an experiment to show that artists are selling music
on jam. With this commit we show the top 8 albums by all-time
purchases, but we might want to scope that to a particular time period
or introuduce some other variation.

To keep rubocop happy I had to add a `dependent: :destroy` option to
the association I added between Album and Purchases, but note that in
purchases in the event that an artist does decide to remove their
albums, so we may need something more sophisticated here in the
future (cc @floehopper).