Add client-side validation for Album#released_on
By adding a `max` attribute with today's date as the value, we can
prevent the album form from being submitted with an invalid (i.e.
future) date.

While this isn't very pretty (because it relies on native browser HTML
form validation), for most users it should prevent the scenario of
losing multiple uploaded tracks if an invalid release date is set.

Fixes #154.