Ensure Time.zone.today is evaluated at comparison time
Before this commit I think `Time.zone.today` was only being evaluated
when the application first booted, rather than each time the
validation was called. This manifested as validation error messages
for some users when the application hadn't been deployed for several
days (e.g. #169).

I'm hopeful that passing a lambda to `less_than_or_equal_to` will fix
this but I'm unsure of how to express that in a test.