Introduce 'new' layout for homepage
I want to replace tailwind with traditional stylesheets, and implement
some of the designs Chad has been working on. To allow me to do that
somewhat incrementally, I'm creating a "new" layout and applying it to
the homepage. At the moment this is a straight copy of the current
application layout.