Use new layout for controllers that redirect to root
Turbo seems to have some issues if the asset bundle changes between
requests. This can happen if, like we've done in the previous commit,
we introduce a new layout. In this case Turbo notices that the asset
bundle has changed between pages and forces a page refresh - this
clears the flash and means flash messages are lost. To avoid that I
need to make sure we use the newly introduced layout on pages that
redirect to the root_path (which also uses the new layout).