Use a reel to show best selling albums
Use the reel layout[1] to avoid so much vertical scrolling on the home
page. I've applied overflow/ellipsis styles to the card headers so
that they don't force the cards to be different widths in the reel. I
also customised the scroll bar colours a little from the book's
original layout.

[1] https://every-layout.dev/layouts/reel/