Add a Cloudflare Turnstill captcha to the signup form
In the Postmark activity I've noticed a lot of "verify your email"
messages being sent and not actioned. Many of these are bouncing,
which hurts our email deliverability. Anecdotally more legitimate
emails from jam.coop are going to spam than before.

To mitigate this I've added a Cloudflare Turnstile[1] "captcha" to the
registration form to hopefully prevent these non-genuine
signups (which I assume are from scripts) from occuring.

I decided to use the rails-cloudflare-turnstile[2] gem to do the heavy
lifting here, primarily to avoid having to write the code that makes
the verfication request to the cloudflare API[3].

I've mostly followed the setup instructions for that gem with a couple
of changes:

1. I've disabled the gem-provided mock in favour of using Cloudflare's
development/testing tokens so that we can see the real capture
locally.
2. Because we use turbo I've had to do ensure that the cloudflare
script is always loaded and reloaded when the signup page is visited
or re-rendered after a validation error. Setting `data-turbo=false` on
the signup form ensures that it is submitted without turbo and
therefore releads the cloudflare JS widget on re-render. Including the
cloudflare script with the `data-turbo-track` and
`data-turbo-temporary` options ensures that it is reloaded if the user
navigates away from signup and back again.

I've also set the two new ENV variables using the credentials provided
by Cloudflare[4].

[1] https://developers.cloudflare.com/turnstile/
[2] https://github.com/instrumentl/rails-cloudflare-turnstile
[3] https://github.com/instrumentl/rails-cloudflare-turnstile/blob/main/lib/rails_cloudflare_turnstile/controller_helpers.rb
[4] https://dash.cloudflare.com/