Add release date to track metadata
Fixes #185

This commit adds the DATE metadata field to the transcoded MP3 and
FLAC files. The FLAC documentation we can find seems to suggest using
DATE for this purpose[1]. For MP3, `date` also seems to be an
acceptable field name[2]. We've verified that the data shows up in the
metadata displayed by VLC[3] for both formats.

[1] https://xiph.org/vorbis/doc/v-comment.html#fieldnames
[2] https://docs.mp3tag.de/format/
[3] https://www.videolan.org/vlc/

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>