Use TORY/DATE tags for release year
Our best understanding is that the MP3 ID3 tag TORY[1] should be used to
store the *year* (i.e. as a 4-digit number) the music was
released. The convention for FLAC tags seems to have even less
consensus/real-world usage, so for simplicity in the code, we've
decided to use the four digit year in the DATE tag.

[1] https://community.mp3tag.de/t/how-do-i-add-the-tag-for-when-my-music-was-released/55319

Co-authored-by: James Mead <james@floehopper.org>