Show recently released albums on homepage
Instead of recently_published. Lots of albums added are from an
Artists back catalogue, but I think here we want to show *new* music
so relying on released on is best. Not all artists set the release
date, but I think that's OK as there's enough that do for this list to
feel fresh.