Run `rails app:update`
I ran the command and reviewed the diff it generated.

I had to add a couple of rubocop exceptions to the migration files to
make them compatible with our cops.

I didn't accept any of the changes to `config/puma.rb` since we use
the suggested puma config provided by render for that.