Use new turbo morph / scroll preserve option
When editing the forms on the user settings page this improves the
experience, as the scroll position is preserved between page
refreshes, making it easier to spot the validation errors if there are
any.

I've had to change the toast message so that it's always fixed in the
viewport on all screen sizes so that it is also always visible when
scroll position is preserved.