Ensure tracks are transcoded on album creation
I seem to have introduced a regression whereby tracks were not
transcoded on album creation. This *might* have something to do with
the `run_after_transaction_callbacks_in_order_defined` behaviour
changed in 935902 affecting the behaviour of the `Album`
`after_commit` hook, but I haven't been able to confirm that.

To fix the issue I'm explictly calling `Album#transcode_tracks` from
the controller on Album create, and I've added a system test to verify
that the Admin user can see all of the transcodes on the Album page.

Closes: #250