Pay user via Stripe Connect during checkout
If the user associated with the artist has successfully setup Stripe Connect on
their account (i.e. it is accepting payments), we supply the account identifier
and the application fee amount to the Stripe checkout session so that the
amount owed to the artist is transferred to their Stripe account immediately.
The user can then control payouts from their Stripe account into their bank
account.

Note that `PurchasesController#create` passes the `StripeConnectAccount`
associated with the user to the `StripeService` if one exists; otherwise
it passes a new instance of `StripeConnectAccount`. This is so that we
can safely call `StripeConnectAccount#connect?` either way, i.e. it's
using the Null Object pattern.

This uses a Stripe Connect Direct Charge [1]:

Previously:
* The charge was made against Jam's Stripe account.
* Jam was responsible for Stripe processing fees.
* Jam was the "Merchant of Record" for the buyer.
* Jam was responsible for taxes, refunds, etc.

Now:
* The charge is made against the seller's Stripe account.
* The seller is responsible for Stripe processing fees.
* The seller is the "Merchant of Record" for the buyer
* The seller is responsible for taxes, refunds, etc.
* Jam deducts an application fee from the amount paid to the seller.

[1]: https://docs.stripe.com/connect/direct-charges