Pay user via Stripe Connect during checkout
If the user associated with the artist has successfully setup Stripe
Connect on their account, we supply the account identifier and the
application fee amount to the Stripe checkout session so that the amount
owed to the artist is transferred to their Stripe account immediately.
The user can then control payouts from their Stripe account into their
bank account.

Note that `PurchasesController#create` passes the `StripeConnectAccount`
associated with the user to the `StripeService` if one exists; otherwise
it passes a new instance of `StripeConnectAccount`. This is so that we
can safely call `StripeConnectAccount#connect?` either way, i.e. it's
using the Null Object pattern.