Keep StripeConnectAccount in sync with Stripe
Stripe sends "account.updated" webhook events when an account's state
changes, so although we check `details_submitted` & `charges_enabled`
properties on the account in `StripeConnectAccountsController#success`,
it's possible they might not have been updated at that point or they
might get updated out-of-band by the user interacting with their Stripe
account. So it seems safer to have this webhook handler in place.