Fix atom feed for artist with no published albums
Previously we were seeing exceptions like this one [1]:

    ActionView::Template::Error: undefined method `released_on' for nil:NilClass

This was because the atom feed was assuming that the artist always had
at least one published album, but it seems as if it's possible to view
the artist page for such an artist which makes the atom feed available,
so we need to make it resilient to that scenario.

This uses `Artist#updated_at` as a fallback for the feed's `updated`
property.

I also had to add the top-level `author` property, because otherwise the
atom feed would be invalid when it has no entries. It seemed simpler to
add it in all scenarios; not just when there were no entries.

[1]: https://app.rollbar.com/a/gofreerange/fix/item/jam/122