Ensure tracks added to published albums are transcoded
Before this commit when an artist added a track to a published album,
the new track was not transcoded, and therefore could not be listened
to.

I've added a system spec to illustrate the bug and also provide some
coverage for editing a published track's title.

I can't quite work out what causes the bug, it seems that the
`after_commit` hook doesn't show any previous changes in the case of a
new track added to the album. Adding an explicit `after_create` fixes
this.

Since we're now generating download zip bundles on purchase, rather
than on creation, we don't have to worry about regenerating those on
this kind of change.

Fixes: #270