Update AWS CDK app to latest release
I installed the latest version (v2.1021.0) of the aws-cdk package and
generated a new app in another directory using the following command:

    cdk init app --language=typescript

I then compared the generated files with those in the aws-cdk directory
of this repo and updated the relevant files accordingly.

Applying these changes against the development & production stacks only
changes the node.js version and the S3 key for the
`Custom::CrossRegionExportReaderCustomResourceProvider/Handler` Lambda
Function which shouldn't cause any problems.