Add purchase_downloads table
We want to change the way downloads are created. We'll create them
after a Purchase is created rather than when an Album is
published. This way we don't have to worry about re-generating them
when the content of an album changes, since the buyer will always
receive the latest version on purchase.

To make this change in smaller steps we first create a duplicate of
the `downloads` table.