Run ZipDownloadJob after Purchase created
Instead of after Album#publish. ZipDownloadJob previously created the
zip files that a user can purchase after an album was published. This
commit changes it so that it can be called after a Purchase is created.

Note that no system specs currently fail, this is because the test
setup in BuyingAnAlbumTest, for example, creates the Download as it
would be during Album publication. I'll change the test (and
implementation) in the subsequent commits to use the new
PurchaseDownload model.