Ensure Album#first_published_on is set when changed
Now we are allowing artists to set the publication status themselves,
we are no longer calling the Album#publish method, so we've lost the
ability to record the first_published_on date. I think it's important
we keep this, since we use it in lieu of a `released_on` date if that
isn't set, and we may later want to use it to decide how to notify
people when an album is first published, even if it's visibility is
subsequently modified.

I've added this behaviour to an after_update hook, so that I can then
remove the Album#publish method.