Ensure newly created published albums have first_published_on
Before this commit if an artist created a new album and selected
"published" as the visibility, the first_published_on was not set,
since we were only setting that in an `after_update` action.

We saw a production issue caused by this[1]. I've fixed the bad data
in production but this change should prevent it from happening again.

https://app.rollbar.com/a/gofreerange/fix/item/jam/128