Update config files
I ran `rails app:update` and then fixed rubocop errors with `rubocop -A`.

I've gone through the proposed file changes and retained any
uncontroversial ones. I've discarded changes to the config/environment
files since the app boots, the tests pass and we have a number of
changes to those specific for this app.