Provide default/placeholder image in CardComponent
We've seen an error in users/show (/account) when we attempt to show
an artist that doesn't have a profile image.

This commit provides a blury default image for `CardComponent`s where
`image` is `nil`. I've used picsum to generate this for now, but we
can replace it with something local later to save the network request.