Add albums atom feed
We want to make it more obvious that music is being added to jam.coop,
as a first step we're introducing this RSS feed that we can
potentially use to automatically post to social media, as well as
allow users to follow activity in their RSS readers. We've also had
one or two people ask for this specific feed to make it easier for
them to write round-ups of new releases.

The format of this RSS feed is closely based on the /artists
feed. We've decided to order this feed by the date the album was added
to jam (rather than the `released_on` date), as that seems more useful
if you want to know when new music has been added.

Co-authored-by: Chris Lowis <chris.lowis@gofreerange.com>
Co-authored-by: James Mead <james.mead@gofreerange.com>