Remove mise.toml configuration file
This is specific to the tool that the developer chooses to manage their
Ruby and Node installs. I think it's better not to have this file in the
repo and instead to configure the relevant tool to respect the generic
`.ruby-version` & `.node-version` files.

To do this in `mise` [1] run the following command:

    mise settings add idiomatic_version_file_enable_tools ruby,node

To do this in `asdf` [2] add the following line to `.asdfrc` [3]:

    legacy_version_file = yes

If you actively want to use `mise.toml` instead then I suggest you add
it to your global git ignores so it's not included in the git repo.

[1]: https://mise.jdx.dev/configuration.html#idiomatic-version-files
[2]: https://asdf-vm.com/manage/configuration.html#legacy-version-file
[3]: https://asdf-vm.com/manage/configuration.html#asdfrc