Handle multiple sessions in take_failed_screenshot
The built-in screenshotting that happens when a system test fails is implemented in
`ActionDispatch::SystemTesting::TestHelpers::ScreenshotHelper#take_failed_screenshot` [1].

However, it only captures a screenshot for the default Capybara session.
We're using multiple sessions in some of our system tests and if a
failure happens in one of those, we get no useful information in the
screenshot.

While it's not ideal that I'm monkey-patching the method or calling a
private `Capybara` method, I think it's worth it in this case, given the
consequences of it breaking are not hugely significant. I've tried to
reproduce as much of the original behaviour as possible, but I haven't
bothered trying to make the in-terminal image stuff work.

[1]: https://api.rubyonrails.org/v8.0.2.1/classes/ActionDispatch/SystemTesting/TestHelpers/ScreenshotHelper.html#method-i-take_failed_screenshot