Reduce cover image size on albums page
To reduce the page load time.

I haven't made a similar change to the albums RSS feed, because (a) the
size is 750x750 which is already smaller than 1000x1000; and (b) RSS
feed readers are likely to only load one entry at a time and thus only
one album cover image at a time.