Limit number of albums to 20 on index page & feed
To reduce the load time for the albums index page and the albums RSS
feed.

We'll probably want to introduce some kind of pagination eventually, but
this will do for now.

RSS feeds don't typically include the entries for all time, because
readers are stateful, so we just need to supply a recent subset.