Limit number of artists to 20 in RSS feed
To reduce the load time for the artists RSS feed. RSS feeds don't
typically include the entries for all time, because readers are
stateful, so we just need to supply a recent subset.

I haven't made a similar change to the artists index page, because (a)
the number of artists isn't currently a problem; and (b) an artist might
be confused or upset if they disappeared off this page!