Script to fetch genres from musicbrainz
Chris and I spent some time discussing how to make music more easily
findable for casual visitors to the website. We've decided to expose
"genres" (e.g. Ambient, Jazz), in a similar way to other online music
services.

We want to pre-populate a list of common genres, and for that we
landed on using the Musicbrainz genre list. This has the advantage
that there is already some modelling in place (e.g. Bebop is a
sub-genre of Jazz[1]). The musicbrainz community also take care of
adding new genres and enriching them with relevant linked data.

As a first step, this commit adds a script to scrape the Musicbrainz
genre-list via its API and save that to a local file we commit to
this repository.

We're using the HTTP Musicbrainz API endpoint to avoid some SSL issues
we encountered when attempting to use their HTTPS endpoint. We've set
a User-Agent header as requested by the Musicbrainz
documentation. We also sleep for 5 seconds between each request to
keep within their 1-request-per-second API limit.

[1] https://musicbrainz.org/genre/ac30756b-55a7-4dc3-8908-8f94c93318d0

Co-authored-by: Chris Roos <chris.roos@gofreerange.com>